==> 3_model_training.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd 
from pickle import load
import os.path as path
from pickle import load, dump

from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier

# load dataframes
X_train_scaled_resamp = pd.read_csv(path.join("3_X_fitted_dataframe", "X_train_scaled_resamp.csv"))
y_train_resamp = pd.read_csv(path.join('4_y_dataframe', 'y_train_resamp.csv'))
X_test_scaled = pd.read_csv(path.join("3_X_fitted_dataframe", "X_test_scaled.csv"))
y_test = pd.read_csv(path.join('4_y_dataframe', 'y_test.csv'))

# We limit the depth of the models intentionnaly to take less time to fit
## We remind that we are not focused on the model accuracy

# fit xgboot model
XGB_clf = XGBClassifier(max_depth=1)
XGB_clf.fit(X_train_scaled_resamp, y_train_resamp)

# fit random forest model
rf_clf = RandomForestClassifier(max_depth=1)
rf_clf.fit(X_train_scaled_resamp, y_train_resamp)

# fit Gradient boostinf model
GBoost_clf = GradientBoostingClassifier(max_depth=1)
GBoost_clf.fit(X_train_scaled_resamp, y_train_resamp)

# save models
dump(rf_clf, open(path.join('6_models', 'randomForest.pkl'), 'wb'))
dump(XGB_clf, open(path.join('6_models', 'xgBoost.pkl'), 'wb'))
dump(GBoost_clf, open(path.join('6_models', 'GradientBoosting.pkl'), 'wb'))
